% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_clines.R
\name{add_clines}
\alias{add_clines}
\title{Adds LaTex \verb{\\clines} to an object of class \code{"xtable"}}
\usage{
add_clines(xtab, rows = "multirow", cols, check.column)
}
\arguments{
\item{xtab}{the \code{xtable} object}

\item{rows}{either \code{"multirow"} (the default) or row numbers to add a cline
to}

\item{cols}{a character string with the columns over which to draw the cline
(e.g, \code{"2-3"})}

\item{check.column}{a character string of the column to check when
\code{rows = "multirow"}}
}
\value{
A list with levels 'pos' and 'command' to give to the \code{add.to.row}
option of \code{xtable()}.
}
\description{
Produces a list with 'pos' and 'command' to use with \code{xtable()}'s
\code{add.to.row} option. This will create a partial horizontal line across
specified columns of the table. Users can either specify the row numbers
they'd like to add the cline to or the function will look for the rows which
use multirow in a specified column.
}
\examples{

library(xtable)

xtab <- xtable(head(iris))

print(xtab, add.to.row = add_clines(., rows = c(2, 3), cols = "3-4"))

}
