% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table.R
\name{ae_table}
\alias{ae_table}
\title{Produce a data frame to summarise count data.}
\usage{
ae_table(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been renamed to \code{\link[=count_table]{count_table()}} to reflect the fact that
it can be used to summarise counts of any events.
}
\keyword{internal}
