% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_to_row.R
\name{colnames_to_row}
\alias{colnames_to_row}
\title{Move the column names of a data frame to a row}
\usage{
colnames_to_row(df, row = 1)
}
\arguments{
\item{df}{the data frame}

\item{row}{the row number before which the column names will be inserted,
default is (i.e., the names would become the first row of the data
frame).}
}
\value{
a data frame with column names in tact but added as a row
}
\description{
Moving the column names of a data frame to a row is sometimes useful when
formatting a table for export. This function does that and the user can
specify which row they'd like the column names to be insterted before.
}
\examples{

colnames_to_row(iris)
colnames_to_row(iris, row = 5)

}
