% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense.R
\name{condense}
\alias{condense}
\title{Condense two or three columns into one}
\usage{
condense(
  df,
  first_col = variable,
  second_col = scoring,
  third_col,
  hline = TRUE,
  indent = c("quad", "hang", "space")
)
}
\arguments{
\item{df}{a data frame containing at least one column with repeated values.}

\item{first_col}{variable name of the column to collapse into - should contain
repeated values. Default is \code{variable} for quick use with the summary table
functions.}

\item{second_col}{variable name of the column to be collapsed into
\code{first_col}. Default is \code{scoring} for quick use with the summary table
functions.}

\item{third_col}{Optionally, a third column to be collapsed into \code{second_col}.}

\item{hline}{Logical indicating whether \code{"\\\\hline\\n"} should be added at the
start of each unique value in \code{first_col}. For use with \code{\link[xtable:xtable]{xtable::xtable()}}.
Default is \code{TRUE}.}

\item{indent}{Character indicating the type of indent to use (the default is
\code{"quad"}):
\itemize{
\item \code{"quad"} uses \eqn{\LaTeX}'s \verb{\\\\quad} to indent the second column (and
\verb{\\\\qquad \\\\quad} to indent the third column if given)
\item \code{"hang"} uses \eqn{\LaTeX}'s \verb{\\\\hangindent2em\\\\hangafter0} to indent the
second column (and \verb{\\\\hangindent4em\\\\hangafter0} to indent the third
column if given)
\item \code{"space"} uses \code{"    "} to indent the second column (and \code{"        "} to
indent the third column if given)
}}
}
\description{
Useful for saving space with summary tables created by \code{\link[=continuous_table]{continuous_table()}},
\code{\link[=discrete_table]{discrete_table()}}, \code{\link[=ticked_table]{ticked_table()}} or \code{\link[=ae_table]{ae_table()}}.
}
\examples{
  continuous_table(df = iris, Petal.Length, Petal.Width, group = Species) \%>\%
       condense()

}
