% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{continuous_table}
\alias{continuous_table}
\title{Produce a data frame to summaries continuous variables}
\usage{
continuous_table(
  df = .,
  ...,
  group,
  time,
  total = TRUE,
  digits = 2,
  condense = FALSE
)
}
\arguments{
\item{df}{Data frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{time}{Optional variable for repeated measures
(currently must me used with group)}

\item{total}{Logical indicating whether a total column should be created}

\item{digits}{Number of digits to the right of the decimal point}

\item{condense}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{condense = TRUE} is
deprecated, use \code{\link[=condense]{condense()}} instead.}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a data frame and produces grouped or un-grouped summaries
such as mean and standard deviation for continuous variables.
}
\examples{
    continuous_table(outcome, score, group = group)
    continuous_table(outcome, score, group = group, condense = FALSE,
                     total = FALSE)
    continuous_table(outcome, score, digits = 1)

}
