% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_table.R
\name{count_table}
\alias{count_table}
\title{Produce a data frame to summarise counts of events}
\usage{
count_table(df, ..., group, ID, N, accuracy = 0.1, total = FALSE)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Variable that defines the grouping}

\item{ID}{Variable that defines the individual identifier (e.g. screening
number)}

\item{N}{a data frame with the group counts (typically produced using
\code{\link[dplyr:count]{dplyr::count()}})}

\item{accuracy}{see details of \code{\link[scales:percent_format]{scales::percent()}}}

\item{total}{Logical indicating whether a total column should be created}
}
\description{
Takes a data frame of events and produces the number of events
and number and percentage of individuals with at least one event.
}
\examples{
  N <- dplyr::count(outcome, group, name = "N")
  count_table(outcome_aes, serious, related, group = group, N = N, ID = screening)

}
