% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagPlot.R
\name{diagPlot}
\alias{diagPlot}
\title{Diagnostic plots}
\usage{
diagPlot(model, ...)
}
\arguments{
\item{model}{a model object for which diagnostic plots are required}

\item{...}{arguments to pass to methods}
}
\value{
returns a \code{gtable}
}
\description{
\code{diagPlot} is a generic function used to produce diagnostic plots for the
results of various model fitting functions. The function invokes particular
\code{\link[=methods]{methods()}} which depend on the \code{\link[=class]{class()}} of the model.
}
