% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{discrete_table}
\alias{discrete_table}
\title{Produce a dataframe to summarise discrete variables}
\usage{
discrete_table(df = ., ..., group = ., total = TRUE)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{total}{Logical indicating wether a total column should be created}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces the number and percentage for
             discrete variables.
}
\examples{
    library(ggplot2) # for the data
    discrete_table(df = mpg, drv, year, group = manufacturer)
    discrete_table(df = mpg, drv)

}
