% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table_r.R
\name{discrete_table_r}
\alias{discrete_table_r}
\title{Produce a dataframe to summarise discrete variables with row percentages.}
\usage{
discrete_table_r(df = ., ..., group = .)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Variable that defines the grouping}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces the number and percentage for
             discrete variables. Unlike \code{discrete_table} this function
             gives row sums.
}
\examples{
    library(ggplot2) # for the data
    discrete_table_r(df = mpg, drv, year, group = manufacturer)

}
