% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicol.R
\name{display_unicol}
\alias{display_unicol}
\alias{get_unicol}
\title{Use the University of Sheffield Colour Palette}
\usage{
display_unicol()

get_unicol(names, tints = 1)
}
\arguments{
\item{names}{a character vector of colour names}

\item{tints}{a real vector of tints, default is 1}
}
\description{
\code{display_unicol()} displays the colour palette in a graphics window.

\code{get_unicol()} returns the hex colour code for specified colour names and
tints.
}
