% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Creates forest plots}
\usage{
forest_plot(df = ., names = x, estimate = y, lower = ymin,
  upper = ymax, xlab = "", ylab = "", title = "", theme = theme_bw(),
  zeroline = TRUE, ...)
}
\arguments{
\item{df}{A data frame containing the names of each analysis as well as the
point estimages and confience interval levels.}

\item{names}{The name of the variable containing the analises names.}

\item{estimate}{The name of the variable containing the point estimates}

\item{lower}{The name of the variable containing the lower limits of the
confidence intervals.}

\item{xlab}{String of text for the x-axis label.}

\item{ylab}{String of text for the y-axis label.}

\item{title}{String of text for the title.}

\item{theme}{ggplot theme. If \code{theme = NULL} then the default ggplot
theme is used.}

\item{zeroline}{Logical. Should a line be added to indicate zero effect?}
}
\description{
Forest plots are often used when comparing different analyses.
             This function uses ggplot2 to create a forest plot.
}
\examples{
df <- data.frame(x = c("A", "B", "C"), y = 1:3, ymin = 0:2, ymax = 2:4)
forest_plot(df)
forest_plot(df, theme = NULL, zeroline = FALSE)

}
