% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_log.R
\name{git_log}
\alias{git_log}
\title{Generates a log of git commits}
\usage{
git_log(filename = file_stamp("git_log", ".csv"))
}
\arguments{
\item{filename}{a character string with the file name to be used, must end in
".csv". The default is "git_log_DATESTAMP.csv".}
}
\description{
This function uses a call to git on the command line to generate
a csv file which logs all git commits within the working directory.
}
