% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longtable_head.R
\name{longtable_head}
\alias{longtable_head}
\title{For use with \code{xtable}'s \code{add.to.row} when
\code{tabular.environment = "longtable"}}
\usage{
longtable_head(
  xtab,
  add.to.caption = " (continued)",
  footnote = "Continued on next page",
  double.header = FALSE,
  pos = 0
)
}
\arguments{
\item{xtab}{An object of class \code{xtable}.}

\item{footnote}{A character string}
}
\value{
A single character string
}
\description{
Format header information from an xtable for adding to the
printed output at position 0.
}
\examples{

xtable::xtable(iris, caption = "Iris Data") \%>\%
    xtable::print.xtable(tabular.environment = "longtable", floating = F,
                         hline.after = c(-1,0),
                         add.to.row = longtable_head(.))

}
