% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Mean (SD)}
\usage{
mean_sd(
  x,
  digits = 2,
  na_rm = FALSE,
  show_n = FALSE,
  note = FALSE,
  unit = NULL
)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits to round mean and standard deviation to}

\item{na_rm}{remove \code{NA} values, default is \code{FALSE}}

\item{show_n}{Should the denominator be shown, default is \code{FALSE}}

\item{note}{should "SD " be written inside the brackets, default is \code{FALSE}}

\item{unit}{optional character string giving unit, e.g., "kg"}
}
\description{
Calculates and formats the mean and standard deviation.
}
\examples{
  mean_sd(outcome$score, na_rm = TRUE)
  mean_sd(outcome$score, na_rm = TRUE, show_n = TRUE)
  mean_sd(outcome$score, na_rm = TRUE, note = TRUE)

}
