% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_iqr.R
\name{median_iqr}
\alias{median_iqr}
\title{Median (IQR)}
\usage{
median_iqr(x, digits = 2, na_rm = FALSE, show_n = FALSE, note = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{number of digits to round mean and standard deviation to}

\item{na_rm}{remove \code{NA} values, default is \code{FALSE}}

\item{show_n}{Should the denominator be shown, default is \code{FALSE}}

\item{note}{should "IQR " be written inside the brackets, default is \code{FALSE}}
}
\description{
Calculates ad formats the median and interquartile range.
}
\examples{
  median_iqr(outcome$score, na_rm = TRUE)
  median_iqr(outcome$score, na_rm = TRUE, show_n = TRUE)
  median_iqr(outcome$score, na_rm = TRUE, note = TRUE)

}
