% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_colnames.R
\name{merge_colnames}
\alias{merge_colnames}
\title{For use with \code{xtable}'s \code{add.to.row} when column names are repeated and
should be merged.}
\usage{
merge_colnames(xtab)
}
\arguments{
\item{xtab}{an object of class \code{xtable}}
}
\value{
a list with components 'pos' and 'command'
}
\description{
Removes consecutive duplicate column names and uses LaTeX's \verb{\\multicolumn} to
merge the cells. Shuld be passed to \code{add.to.row} in \code{print.xtable()}.
}
\examples{

library(xtable)

iris2 <- split_colnames(iris, sep = "\\\\.")

xtable(iris2) \%>\%
  print.xtable(add.to.row = merge_colnames(.))

}
