% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_select_filter.R
\name{merge_select_filter}
\alias{merge_select_filter}
\title{Combines the functions merge, select and filter to simplify combining
data.frames from prospect.}
\usage{
merge_select_filter(data.x = ., data.y = ., select.x = c(),
  select.y = c(), filter.x = ., filter.y = ., by = c(), ...)
}
\arguments{
\item{data.x}{The first data frame.}

\item{data.y}{The second data frame.}

\item{select.x}{Variables to select from the first data frame.}

\item{select.y}{See \code{select.x}.}

\item{filter.x}{Rules for filtering the first data frame, unlike in the
function \code{filter()} we must use \code{&} to sepperate
our rules.}

\item{filter.y}{See \code{filter.x}.}

\item{by}{A vector of character strings giving the columns to merge by.}
}
\value{
A tibble data frame combining the two data frames with the rules
        given applied.
}
\examples{
   x <- data.frame(ID = 1:5, height = rnorm(5, 167, 20),
                   group = sample(c("A", "B"), 5, replace = T))
   y <- data.frame(ID = 1:5, weight = rnorm(5, 70, 15),
                   waist = rnorm(5, 32, 3))

   merge_select_filter(x, y, select.x = c(ID, height),
                       select.y = c(ID, weight),
                       filter.x = height > 130 & group == "A",
                       filter.y = weight > 60, by = "ID")

}
