% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_table.R
\name{missing_table}
\alias{missing_table}
\title{Produce a dataframe to summarise data completeness for variables}
\usage{
missing_table(df = ., ..., group = ., format = "Missing", total = TRUE)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{format}{Should the propotion missing or present be given?}

\item{total}{Logical indicating whether a total column should be created}
}
\value{
A tibble data frame summarising the data completeness
}
\description{
Takes a dataframe and calculates the proportions present/missing
             for given variables.
}
