% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_tidy.emmGrid.R
\name{my_tidy.emmGrid}
\alias{my_tidy.emmGrid}
\title{Extend broom's tidy.emmGrid function to provide confidence intervals}
\usage{
my_tidy.emmGrid(x, conf.int = F, conf.level = 0.95)
}
\arguments{
\item{x}{"emmGrid" object}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence interval if
conf.int = TRUE. Must be strictly greater than 0 and less
than 1. Defaults to 0.95, which corresponds to a 95 percent
confidence interval.}
}
\description{
provide a tidy table for emmGrid objects
}
\examples{
library(emmeans)
data("oranges")

oranges.lm <- lm(sales1 ~ price1*day, data = oranges)
oranges.em <- emmeans(oranges.lm, pairwise ~ day)

my_tidy.emmGrid(oranges.em$contrasts, conf.int = T)

}
