% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcome_aes}
\alias{outcome_aes}
\title{Simulated AE data.}
\format{
A data frame with 100 rows and 3 columns:
\describe{
\item{screening}{character, individual id number}
\item{group}{factor, treatment group A/B}
\item{serious}{factor with two levels: No, Yes}
\item{related}{factor with two levels: No, Yes}
}
}
\usage{
outcome_aes
}
\description{
A simulated data frame used in examples.
}
\keyword{datasets}
