% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pn.R
\name{pn}
\alias{pn}
\alias{np}
\title{Count and Percentage}
\usage{
pn(x, digits = 1, na_rm = FALSE, show_denom = FALSE, paren = "(", note = FALSE)

np(x, digits = 1, na_rm = FALSE, show_denom = FALSE, paren = "(")
}
\arguments{
\item{x}{A boolean vector}

\item{digits}{Number of digits to display percentages to, default is 1}

\item{na_rm}{Logical indicating whether NAs should be removed from the
calculation of the denominator, default is \code{FALSE}}

\item{show_denom}{Logical, should the denominator be shown, default is \code{FALSE}}

\item{paren}{String indicating which parentheses to use, options are \code{"("} (for ()) or \code{"["} (for[])}

\item{note}{Should "n = " preface the count in \code{pn()}}
}
\description{
Functions to display counts and percentages.
}
\examples{
  pn(outcome$limp_yn == "Yes", note = TRUE)
  np(outcome$limp_yn == "Yes", show_denom = TRUE, na_rm = T)


}
