% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevent_pagebreak.R
\name{prevent_pagebreak}
\alias{prevent_pagebreak}
\title{Preventing page breaks when using in long tables exported from \code{xtable()}}
\usage{
prevent_pagebreak(pxtab, lines)
}
\arguments{
\item{pxtab}{the output from a call to \code{print.xtable}}

\item{lines}{either a numeric vector of row numbers or \code{"multirow"}. If the latter then the function will check which lines start with a cell which uses \verb{\\multirow}.}
}
\description{
When using LaTeX's longtable environment it is sometimes desirable to prevent a page break for certain lines in the table. In LaTeX this is done by using \code{*} at the end of the row. \code{prevent_pagebreak} adds \code{*} to specified lines of output from \code{print.xtable}.
}
