% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prospect.R
\name{read_prospect}
\alias{read_prospect}
\title{Reads in csv files as exported by Prospect}
\usage{
read_prospect(file, dictionary = lookups, convert.date = TRUE, ...)
}
\arguments{
\item{file}{The file to be read in}

\item{dictionary}{The file in which the lookups reside, default is
\code{lookups}}

\item{convert.date}{Convert fields ending in \verb{_dt} or \verb{_date} to date format}

\item{...}{arguments to be passed to \code{\link[utils:read.table]{utils::read.csv()}}. \code{stringsAsFactors}
is set to \code{FALSE} within \code{read_prospect()}, re-specifying it in the call
will throw an error.}
}
\value{
An object of class \code{\link{data.frame}}
}
\description{
Takes a csv file exported from Prospect and reads it in to R
along with the factor labels from the \code{lookups.csv} file produced by
Prospect. When using this function you must first read the \code{lookups.csv}
file into R.
}
