% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{For use with tables}
\usage{
remove_duplicates(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
A character string/vector
}
\description{
Turns duplicate rows into NAs
}
\examples{
    x <- c(rep("a", 5), rep("c", 2), rep("y", 7))
    remove_duplicates(x)

}
