% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round}
\alias{round}
\title{Rounds numners but keeps trailing zeros}
\usage{
round(x, digits)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{an integer indicating the number of decimal places}
}
\value{
a character string
}
\description{
This version of \code{round} is a shortcut for
             \code{formatC(x, digits, format = "f")}.
}
\examples{
base::round(5.601, 2)
Useful.functions::round(5.601, 2)

}
