% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Rounds numbers but keeps trailing zeros}
\usage{
round0(x, digits)
}
\arguments{
\item{x}{a numeric vector}

\item{digits}{an integer indicating the number of decimal places}
}
\value{
a character string
}
\description{
This function is a shortcut for
\code{formatC(x, digits, format = "f")}.
}
\examples{
round(5.601, 2)
round0(5.601, 2)

}
