% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_to_colnames.R
\name{row_to_colnames}
\alias{row_to_colnames}
\title{Tools for working with column names}
\usage{
row_to_colnames(df, row = 1)
}
\arguments{
\item{df}{A data frame.}

\item{row}{Row number to convert to the column names, default is the first
row.}
}
\description{
Takes a row of data and uses it to replace the column names.
}
