% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_percent.R
\name{sanitise_percent}
\alias{sanitise_percent}
\title{Sanitise function for use with print.xtable}
\usage{
sanitise_percent(str)
}
\arguments{
\item{str}{String of text}
}
\description{
Takes a string of text and sanitises only the use of "%" for
             use with latex
}
\examples{
    str <- "\\\\multirow{2}{4cm}{75\%}"
    sanitise_percent(str)

}
