% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_latex.R
\name{save_latex}
\alias{save_latex}
\title{Save output from print.xtable}
\usage{
save_latex(latex, file)
}
\arguments{
\item{latex}{Output from \code{print.xtable}}

\item{file}{File location to save to.}
}
\description{
Saves output from print.xtable without having to wrap
             \code{sink()} above and below.
}
