% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_colnames.R
\name{split_colnames}
\alias{split_colnames}
\title{Tools for working with column names}
\usage{
split_colnames(df, sep = "_")
}
\arguments{
\item{df}{A data frame.}

\item{sep}{Separator between column names and first row.}
}
\description{
Splits column names into two rows. The first row are the new
             column names, the second are the first row of the data.
}
