% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strwrap_factor.R
\name{strwrap_factor}
\alias{strwrap_factor}
\title{String wrapping for factor variables}
\usage{
strwrap_factor(x, width, ...)
}
\arguments{
\item{x}{a factor vector}

\item{width}{Positive integer giving target line width (in number of
characters). A width less than or equal to 1 will put each word
on its own line.}

\item{...}{arguments passed to \code{stringr::str_wrap}}
}
\value{
a factor vector the same length as \code{x} with factor levels wrapped
}
\description{
String wrapping for factor variables
}
\examples{

nt_yorks <- factor(c("Beningbrough Hall (Historic House)",
                     "Braithwaite Hall (Historic Property)",
                     "Brimham Rocks (Countryside)",
                     "East Riddlesden Hall (Historic House)",
                     "Fountains Abbey (Abbey)",
                     "Gibson's Mill (Historic Property)",
                     "Goddards House & Garden (Historic House)",
                     "Hardcastle Crags (Countryside)",
                     "Hudswell Woods (Countryside)",
                     "Maister House (Historic Property)",
                     "Moulton Hall (Historic Property)",
                     "Nostell Priory (Historic House)",
                     "Nunnington Hall (Historic Property)",
                     "Rievaulx Terrace and Temples (Garden)",
                     "Roseberry Topping (Countryside)",
                     "Studley Royal Water Garden (Garden)",
                     "The Bridestones (Countryside)",
                     "Treasurers House (Historic Property)"))

strwrap_factor(nt_yorks, width = 10)

}
