% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_with_footnotes.R
\name{table_with_footnotes}
\alias{table_with_footnotes}
\title{Wrap for print.xtable to add threeparttable functionality}
\usage{
table_with_footnotes(df, align, include.rownames = FALSE,
  include.colnames = FALSE, table.pos = "h!", footnotes = list(i, text),
  caption, label, ...)
}
\arguments{
\item{df}{Data frame to be passed to \code{print(xtable())}}

\item{align}{Character string specifying the column alighnment. See
code{?xtable} for more information.}

\item{include.rownames}{Logical indicating whether the row names of
\code{df} are to be included in the table}

\item{include.colnames}{Logical indicating whether the column names of
\code{df} are to be included in the table}

\item{table.pos}{Character string indicating the position to pass to the
floating environment which is \code{table}}

\item{footnotes}{A list containing two components. The first should be
called "i" and should contain a character vector with the
names of the footnotes in the table (e.g.
\code{c("\\\\dag")}). The second component should be called
"text" and should be a character vector of equal length to
"i". "text" should contain the footnote text to be
displayed below the table.}

\item{caption}{Optional character string giving the caption for the table}

\item{label}{Optional character string giving the label for the table}

\item{...}{Further arguments to be passed to \code{print.xtable}}
}
\description{
This function wraps \code{print.xtable} with the threeparttable
             environment from latex.
}
\note{
The package threeparttable is required to run the produced code in
      latex
}
\examples{
df <- BOD
df[1,1] <- paste(df[1,1], "\\\\tnote{1}")
df[4,2] <- paste(df[4,2], "\\\\tnote{2}")
table_with_footnotes(df,
                     align = "lrr",
                     footnotes = list(i = c(1,2), text = c("a", "b"),
                     hline.after = c(1,2),
                     caption = "Yes",
                     label = "tab:yes")

}
