% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totals.R
\name{totals}
\alias{totals}
\title{Mutates data frame so that summary tables give totals}
\usage{
totals(df = ., ..., name = "Total")
}
\arguments{
\item{df}{A data frame}

\item{...}{one or two columns for which we require totals}

\item{name}{Character string for the name of the total row/column}
}
\description{
This function uses \code{stat_expand} and \code{mutate} to give
             a data frame that, when summarised, with give total rows and/or
             columns.
}
\examples{
    totals(mtcars, cyl) \%>\%
    count(cyl)

    totals(mtcars, cyl, vs, name = "All") \%>\%
    count(cyl, vs) \%>\%
    spread(cyl, n)

}
