% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unicol}
\alias{unicol}
\title{University of Sheffield Colour Palette}
\format{
\subsection{\code{unicol}}{

A data frame with 48 rows and 3 columns:
\describe{
\item{Colour.name}{Colour name}
\item{Tint}{Where, applicable the tint of the colour.}
\item{Hex}{Hex colour code}
}
}
}
\source{
\url{https://www.sheffield.ac.uk/brand-toolkit/colour}
}
\usage{
unicol
}
\description{
A data frame of the colours in the official Univerisity of Sheffield colour
palette.
}
\keyword{datasets}
