% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Useful.functions-package.R
\docType{package}
\name{Useful.functions-package}
\alias{Useful.functions}
\alias{Useful.functions-package}
\title{Useful.functions: Some Useful Functions}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Esther Herbert \email{E.Herbert@sheffield.ac.uk}

}
\keyword{internal}
