% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_lookups.R
\name{wrangle_lookups}
\alias{wrangle_lookups}
\title{Wrangle the Edith database spec into a sing lookups data frame}
\usage{
wrangle_lookups(lookups, fields)
}
\arguments{
\item{lookups}{a data frame read in from the Lookups.csv file as exported
from Edith}

\item{fields}{a data frame read in from the Fields.csv file as exported from
Edith}
}
\description{
This function takes the Lookups and Fields tabs from the Edith database
specification (exported as separate csv files) and wrangles them into a
single lookups data frame for use with \code{\link[=read_prospect]{read_prospect()}} and
\code{\link[=apply_labels]{apply_labels()}}.
}
