% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yn_table.R
\name{yn_table}
\alias{yn_table}
\title{Produce a data frame to summarise discrete variables coded as Yes/No}
\usage{
yn_table(df = ., ..., group, digits = 1, total = TRUE, show_denom = TRUE)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{digits}{Number of digits to display percentages to, default is 1}

\item{total}{Logical indicating whether a total column should be created}

\item{show_denom}{Logical, should the denominator for each variable be shown.
Default is \code{TRUE}.}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a data frame and produces the number and percentage for yes/no
variables. Note that the denominator currently always excludes missing
values.
}
\examples{
yn_table(outcome, limp_yn, group = group)
yn_table(outcome, limp_yn, group = group, total = FALSE)
yn_table(outcome, limp_yn, show_denom = FALSE)


}
