% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table.R
\name{ae_table}
\alias{ae_table}
\title{Produce a dataframe to summarise adverse event data.}
\usage{
ae_table(df = ., ..., group = ., ID = ., N = ., accuracy = 0.1)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Variable that defines the grouping}

\item{ID}{Variable that defines the individual identifier (e.g. screening
number)}

\item{accuracy}{see details of \code{scales::percent}}
}
\description{
Takes a dataframe of adverse events and produces the number of
             events and number and percentage of individuals with an adverse
             event.
}
