% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_table.R
\name{ae_table}
\alias{ae_table}
\title{Produce a data frame to summarise adverse event data.}
\usage{
ae_table(df, ..., group, ID, N, accuracy = 0.1, total = FALSE)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Variable that defines the grouping}

\item{ID}{Variable that defines the individual identifier (e.g. screening
number)}

\item{N}{a data frame with the group counts (typically produced using
\code{dplyr::count})}

\item{accuracy}{see details of \code{scales::percent}}

\item{total}{Logical indicating whether a total column should be created}
}
\description{
Takes a data frame of adverse events and produces the number of
events and number and percentage of individuals with an adverse
event.
}
