% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_labels.R
\name{apply_labels}
\alias{apply_labels}
\title{Apply labels from a dataspec to a column}
\usage{
apply_labels(df, col, dataspec = dataspec)
}
\arguments{
\item{df}{Data frame}

\item{col}{Variable to be labeled}

\item{dataspec}{where to find the labels}
}
\description{
Apply labels from a dataspec to a column
}
