% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{continuous_table}
\alias{continuous_table}
\title{Produce a dataframe to summaries continuous variables}
\usage{
continuous_table(df = ., group = group, variables = c(), total = TRUE,
  set = ., ...)
}
\arguments{
\item{df}{Data frame}

\item{group}{Variable that defines the grouping}

\item{variables}{Variables to be summarised}

\item{total}{logical indicating wether a total column should be created}

\item{set}{Optional variable that defines extra grouping. This is grouping
obove the grouping given in \code{group}}
}
\description{
Takes a dataframe and produces grouped summaries such as mean
             and standard deviation for continuous variables.
}
\examples{
    continuous_table(df = iris, group = Species,
                     variables = c(Petal.Length, Petal.Width))
    continuous_table(df = iris, group = Species,
                     variables = c(Sepal.Length, Sepal.Width),
                     total = FALSE)

}
