% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{continuous_table}
\alias{continuous_table}
\title{Produce a dataframe to summaries continuous variables}
\usage{
continuous_table(df = ., ..., group = ., total = TRUE, digits = 2)
}
\arguments{
\item{df}{Data frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{total}{Logical indicating wether a total column should be created}

\item{digits}{Number of digits to the right of the decimal point}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces grouped or ungrouped summaries
             such as mean and standard deviation for continuous variables.
}
\examples{
    continuous_table(df = iris, Petal.Length, Petal.Width, group = Species)
    continuous_table(df = iris, Sepal.Length, Sepal.Width, group = Species,
                     total = FALSE)
    continuous_table(df = iris, Petal.Length, Sepal.Length, digits = 1)

}
