% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_table.R
\name{continuous_table}
\alias{continuous_table}
\title{Produce a dataframe to summaries continuous variables}
\usage{
continuous_table(df = ., group = group, variables = c(), ...)
}
\arguments{
\item{df}{Data frame}

\item{group}{Variable that defines the grouping}

\item{variables}{Variables to be summarised}
}
\description{
Takes a dataframe and produces grouped summaries such as mean and
             standard deviation for continuous variables.
}
