% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagPlot.R
\name{diagPlot}
\alias{diagPlot}
\title{Diagnostic plots}
\usage{
diagPlot(model, ...)
}
\arguments{
\item{model}{a model object for which diagnostic plots are required}

\item{plot}{logical indicating whether the results should be plotted
immediately}
}
\value{
returns a `gtable`
}
\description{
`diagPlot` is a generic function used to produce diagnostic plots for the
results of various model fitting functions. The function invokes particular
[methods()] which depend on the [class()] of the model.
}
