% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{discrete_table}
\alias{discrete_table}
\title{Produce a dataframe to summarise discrete variables}
\usage{
discrete_table(df = ., group = group, variables = c(), total = TRUE,
  set = ., ...)
}
\arguments{
\item{df}{Data Frame}

\item{group}{Variable that defines the grouping}

\item{variables}{Variables to be summarised}

\item{total}{Logical indicating wether a total column should be created}

\item{set}{Optional variable that defines extra grouping. This is grouping
obove the grouping given in \code{group}}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces the number and percentage for
             discrete variables.
}
\examples{
    library(ggplot2)
    discrete_table(df = mpg, group = manufacturer, variables = c(drv, year))

}
