% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{discrete_table}
\alias{discrete_table}
\title{Produce a dataframe to summarise discrete variables}
\usage{
discrete_table(df = ., ..., group = ., total = TRUE, n = FALSE, accuracy = 0.1)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{total}{Logical indicating whether a total column should be created}

\item{n}{Logical indicating whether percentages should be out of n
(\code{n = TRUE}) or N (\code{n = FALSE})}

\item{accuracy}{see details of \code{scales::percent}}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces the number and percentage for
             discrete variables.
}
\examples{
    library(ggplot2) # for the data
    discrete_table(df = mpg, drv, year, group = manufacturer)
    discrete_table(df = mpg, drv)

}
