% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{discrete_table}
\alias{discrete_table}
\title{Produce a dataframe to summarise discrete variables}
\usage{
discrete_table(df = ., group = group, variables = c(), ...)
}
\arguments{
\item{df}{Data Frame}

\item{group}{Variable that defines the grouping}

\item{variables}{Variables to be summarised}
}
\description{
Takes a dataframe and produces the number and percentage for
             discrete variables.
}
