% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_table.R
\name{discrete_table}
\alias{discrete_table}
\title{Produce a dataframe to summarise discrete variables}
\usage{
discrete_table(
  df = .,
  ...,
  group = .,
  time = .,
  total = TRUE,
  n = FALSE,
  missing = "Missing",
  accuracy = 0.1,
  condense = FALSE
)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{time}{Optional variable for repeated measures
(currenlty must me used with group)}

\item{total}{Logical indicating whether a total column should be created}

\item{n}{Logical indicating whether percentages should be out of n
(\code{n = TRUE}) or N (\code{n = FALSE})}

\item{missing}{String determining what missing data will be called
(if \code{n = TRUE}). Default is "Missing".}

\item{accuracy}{see details of \code{scales::percent}}

\item{condense}{should the \code{variable} and \code{scoring} columns in the output be
condensed to one column?}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces the number and percentage for
discrete variables.
}
\examples{
    library(ggplot2) # for the data
    discrete_table(df = mpg, drv, group = manufacturer)
    discrete_table(df = mpg, drv)

}
