% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_check.R
\name{export_check}
\alias{export_check}
\title{Checks imported form lengths against the export note produced by prospect.}
\usage{
export_check(export_notes, data)
}
\arguments{
\item{export_notes}{file path to the export note (usually in the same folder
as the exported csv files and called export_note.txt)}

\item{data}{a list containing the read in data}
}
\value{
nothing is returned but a message is printed to the console if the
         check is successful and an error is given if not.
}
\description{
For use with CTRU studies using prospect. This function checks the read in
csv file lengths against he export note produced by prospect.
}
