% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_stamp.R
\name{file_stamp}
\alias{file_stamp}
\title{Produce a file name string with date-time stamp}
\usage{
file_stamp(chr, format = ".xlsx", time = FALSE)
}
\arguments{
\item{chr}{String detailing file name}

\item{format}{String giving file extension}

\item{time}{Logical - should time be included in the stamp?}
}
\value{
A character string with the new file name
}
\description{
Turns a character string into a file name with date-time stamp
}
