% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outcome}
\alias{outcome}
\title{Simulated outcome data.}
\format{
\subsection{\code{outcome}}{

A data frame with 600 rows and 6 columns:
\describe{
\item{site}{character, site code}
\item{screening}{character, individual id number}
\item{group}{factor, treatment group A/B}
\item{sex}{factor with three levels: Male, Female, Prefer not to specify}
\item{event_name}{factor with 3 levels: Baseline, 6 Weeks, 12 Weeks}
\item{score}{double, randomly generated "score"}
}
}
}
\usage{
outcome
}
\description{
A simulated data frame used in examples.
}
\keyword{datasets}
