% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prospect.R
\name{read_prospect}
\alias{read_prospect}
\title{Reads in csv files as exported by Prospect}
\usage{
read_prospect(file = ., dictionary = lookups, convert.date = TRUE,
  ...)
}
\arguments{
\item{file}{The file to be read in}

\item{dictionary}{The file in which the lookups reside, default is \code{lookups}}

\item{convert.date}{Convert fields ending in \code{_dt} to date format}
}
\value{
A data frame formatted as required
}
\description{
Takes a .csv file exported from Prospect and reads it in to R
             along with the factor labels from the looups.csv file produced
             by Prospect. When using this function you must first read the
             lookups.csv file into R.
}
