% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_list.R
\name{search_list}
\alias{search_list}
\title{Searches a list of dataframes.}
\usage{
search_list(ls, string, exact = TRUE)
}
\arguments{
\item{ls}{A list of dataframes}

\item{string}{A string, either the exact variable name or something to search
for.}

\item{exact}{Logical. If \code{exact = TRUE} then \code{search_list} finds
the exact variable, if \code{exact = FALSE} then
\code{search_list} finds all variable names containing that
string using \code{stringr::str_detect()}.}
}
\value{
A data.frame
}
\description{
Takes a list of dataframes and searches their variables for
             either a specific variable name or for a string within the
             variable names.
}
\examples{
files <- list(mtcars = mtcars, iris = iris)
search_list(files, "Sepal", exact = F)
search_list(files, "hp")

}
