% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_analysis.R
\name{start_analysis}
\alias{start_analysis}
\title{Start Analysis Template}
\usage{
start_analysis(path)
}
\arguments{
\item{path}{The path for the analysis folder.}
}
\description{
Create a new RStudio project for analysis. Creates a folder
structure and RStudio project.
}
\examples{
try(start_analysis("path/to/new/project"))
}
