% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticked_table.R
\name{ticked_table}
\alias{ticked_table}
\title{Produce a dataframe to summarise ticked variables}
\usage{
ticked_table(
  df = .,
  ...,
  group,
  sep,
  digits = 1,
  total = TRUE,
  n = FALSE,
  boolean = FALSE,
  condense = FALSE
)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{sep}{Optional separator between columns for splitting variable into
variable and scoring. See ?tidyr::separate for more information.}

\item{digits}{Number of digits to the right of the decimal point}

\item{total}{Logical indicating whether a total column should be created}

\item{n}{Logical indicating whether the second header row should be converted
into a row of denominators. Helpful for when \code{ticked_table()} forms part of
a larger table. Default is \code{FALSE}}

\item{boolean}{Logical indicating whether the variables are already coded as
TRUE/FALSE or 0/1. Default is \code{FALSE}}

\item{condense}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{condense = TRUE} is
deprecated, use \code{\link[=condense]{condense()}} instead.}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces the number and percentage for
ticked variables.
}
\examples{
  ticked_table(outcome, pet_cat, pet_dog, group = group, sep = "_")

  ticked_table(outcome, pet_fish, pet_dog)

  ticked_table(outcome, pet_fish, pet_dog, sep = "_", n = TRUE)

}
