% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticked_table.R
\name{ticked_table}
\alias{ticked_table}
\title{Produce a dataframe to summarise ticked variables}
\usage{
ticked_table(df = ., group = group, variables = c(), total = TRUE,
  set = ., ...)
}
\arguments{
\item{df}{Data Frame}

\item{group}{Variable that defines the grouping}

\item{variables}{Variables to be summarised}

\item{total}{logical indicating wether a total column should be created}

\item{set}{Optional variable that defines extra grouping. This is grouping
obove the grouping given in \code{group}}
}
\description{
Takes a dataframe and produces the number and percentage for
             ticked variables.
}
