% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticked_table.R
\name{ticked_table}
\alias{ticked_table}
\title{Produce a dataframe to summarise ticked variables}
\usage{
ticked_table(
  df = .,
  ...,
  group = .,
  sep = .,
  digits = 1,
  condense = FALSE,
  total = TRUE
)
}
\arguments{
\item{df}{Data Frame}

\item{...}{Variables to be summarised}

\item{group}{Optional variable that defines the grouping}

\item{sep}{Optional separator between columns for splitting variable into
variable and scoring. See ?tidyr::separate for more information.}

\item{digits}{Number of digits to the right of the decimal point}

\item{condense}{should the \code{variable} and \code{scoring} columns in the output be
condensed to one column? Only used when \code{sep} has been
specified.}

\item{total}{Logical indicating whether a total column should be created}
}
\value{
A tibble data frame summarising the data
}
\description{
Takes a dataframe and produces the number and percentage for
ticked variables.
}
\examples{
df <- data.frame(
         pet_cat = sample(c("Ticked", ""), size = 100, replace = TRUE),
         pet_dog = sample(c("Ticked", ""), size = 100, replace = TRUE),
         pet_pig = sample(c("Ticked", ""), size = 100, replace = TRUE),
         group = sample(c("A", "B", "C"), size = 100, replace = TRUE)
 )

  ticked_table(df, pet_cat, pet_dog, group = group, sep = "_")

  ticked_table(df, pet_pig, pet_dog)

}
